requirements.txt
# Core dependencies
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
alembic==1.13.1

# Vector database
pgvector==0.2.5

# Web scraping and processing
langchain==0.1.10
langchain-together==0.1.0
firecrawl-py==0.0.16
beautifulsoup4==4.12.3
requests==2.31.0

# Machine Learning
sentence-transformers==2.2.2
numpy==1.24.4

# Task scheduling
apscheduler==3.10.4

# Environment and configuration
python-dotenv==1.0.0

# Web framework for viewer
fastapi==0.108.0
uvicorn==0.25.0
jinja2==3.1.3

# Terminal UI and monitoring
rich==13.7.0
asyncio-mqtt==0.16.1

# Utilities
pydantic==2.5.3
loguru==0.7.2
asyncio==3.4.3
